refactor(pulp-react): split prop-applier switch by domain (P5-NEW-A)
The monolithic `applyOne(id, type, key, value, props?)` in
prop-applier.ts carried a ~1,100-line `switch (key)` covering every
React prop the bridge understands. P5-NEW-A splits it by property
domain into focused sibling modules, leaving `applyOne` a thin
dispatcher that calls each handler in sequence until one claims the
key.

The switch was verified mechanically partitionable first: every case
ends in return (only adjacent same-domain label-stacking, no
cross-domain fallthrough), and no mutable local state is shared or
accumulated across cases — each case scopes its own locals.

New modules (returning whether they handled the key):
  - prop-applier-internal.ts   (128) — shared `call` + `_resolveVar`
  - prop-applier-layout.ts     (479) — flex / grid / box-model / position
  - prop-applier-paint.ts      (482) — background / border / shadow / filter
  - prop-applier-typography.ts (228) — text / font / decoration / writing
  - prop-applier-transform.ts  (313) — transform / transition / animation
  - prop-applier-events.ts      (67) — overlay-claim / ARIA interaction

prop-applier.ts: 1,994 -> 517 lines. The `call` helper moves to the
shared internal module so the per-domain handlers share one logging
counter, exactly as before. Domain-specific file-local helpers move
with their domain (e.g. _coerceLen with layout, _parseBoxShadow with
paint, _walkTransformArray with transform). Type-dispatched widget /
SVG props (data / level / value / d / viewBox / fill / stroke /
strokeWidth) stay inline in the dispatcher — they route on widget
`type`, not purely on `key`.

Behavior is byte-identical: all 201 switch case keys preserved, the
five domain modules are key-disjoint, and the 479-test pulp-react
suite passes unchanged.
