+++ packages/pulp-react/src/prop-applier.ts
@@ -10,28 +10,25 @@
 
 import type { PulpInstance } from './types.js';
 import { makeSyntheticEvent } from './synthetic-event.js';
+import { call } from './prop-applier-internal.js';
+import { applyLayoutProp } from './prop-applier-layout.js';
+import { applyPaintProp } from './prop-applier-paint.js';
+import { applyTypographyProp } from './prop-applier-typography.js';
+import { applyTransformProp } from './prop-applier-transform.js';
+import { applyEventProp } from './prop-applier-events.js';
 
-// Bridge globals are looked up through globalThis at call time so the
-// mock-bridge install path picks them up. See host-config.ts for the
-// matching pattern.
+// P5-NEW-A — the former monolithic `applyOne` switch is split into
+// per-domain handler modules (prop-applier-layout / -paint /
+// -typography / -transform / -events). `applyOne` below is now a thin
+// dispatcher that calls each in sequence until one claims the key.
+// The bridge `call` helper and the shared `_resolveVar` resolver live
+// in prop-applier-internal.ts so the domain modules share one logging
+// counter — exactly as it was when `call` lived in this file.
+//
+// Bridge globals are still looked up through globalThis at call time
+// so the mock-bridge install path picks them up (see host-config.ts).
 type AnyFn = (...args: unknown[]) => unknown;
 const g = globalThis as unknown as Record<string, AnyFn | undefined>;
-let _pa_count = 0;
-function call(name: string, ...args: unknown[]): void {
-    const fn = g[name];
-    if (typeof fn !== 'function') return; // optional bridge fns are fine to skip
-    _pa_count++;
-    if (_pa_count <= 100) {
-        const lg = (g as Record<string, AnyFn | undefined>).__spectrLog;
-        if (typeof lg === 'function') {
-            const a0 = args[0] !== undefined ? String(args[0]).slice(0, 25) : '';
-            const a1 = args[1] !== undefined ? String(args[1]).slice(0, 25) : '';
-            const a2 = args[2] !== undefined ? String(args[2]).slice(0, 25) : '';
-            lg('[pa#' + _pa_count + '] ' + name + '(' + a0 + ',' + a1 + (args.length > 2 ? ',' + a2 : '') + ')');
-        }
-    }
-    fn(...args);
-}
 
 let _aap_count = 0;
 function logApply(stage: string, id: string, type: string, propCount: number): void {
@@ -113,446 +110,6 @@ function isWheelEvent(eventName: string): boolean {
     return eventName === 'wheel';
 }
 
-// pulp #1899 (gap #3) — resolve `var(--name [, fallback])` references in
-// string-valued style props before forwarding to the bridge.
-//
-// The HTML-shim path (`core/view/js/css-parser.js`) already resolves
-// var() via the same lookup tiers, but the React-prop-applier lane
-// bypasses that — it forwards the raw `style.fontFamily` value straight
-// into `setFontFamily`. The string "var(--mono)" then reached Skia's
-// font matcher as a literal family name, returned no match, and fell
-// through to a proportional sans (visible as the Spectr top-bar "faint
-// label" symptom — labels rendered in the wrong typeface AND in a layer
-// that degraded the LCD AA, compounding the faintness).
-//
-// Resolution tiers (first hit wins):
-//   1. `globalThis.__pulpCssVars[name]` — developer-set runtime
-//       registry. Apps populate this at mount when they have direct
-//       knowledge of which CSS variables map to which strings.
-//   2. `getStringToken(name)` — bridge call into `theme.strings` (the
-//       same map design-import.cpp writes when loading Stitch / W3C
-//       tokens, and the same map `setStringToken` from the CSS shim
-//       writes for string-valued custom properties).
-//   3. `getMotionToken(name)` — bridge call into `theme.dimensions`
-//       (only useful when the var is numeric — fontFamily callers
-//       won't hit this branch, but length-typed callers might).
-//   4. The literal fallback supplied as `var(--name, FALLBACK)`.
-//   5. The original string, returned unchanged so the caller can decide
-//       how to handle the miss.
-//
-// The regex match is conservative: only top-level `^var(...)` is
-// resolved. Embedded `calc(var(...) + 10px)` is out of scope for the
-// React lane — the bridge consumers for `fontFamily` / `color` /
-// `borderColor` etc. don't accept calc-expressions anyway.
-function _resolveVar(value: unknown): unknown {
-    if (typeof value !== 'string') return value;
-    const s = value.trim();
-    if (!s.startsWith('var(') || !s.endsWith(')')) return value;
-
-    // Find the FIRST top-level comma inside var(...) so a nested
-    // `var(--a, var(--b, default))` is split at the right point.
-    const inner = s.slice(4, s.length - 1);
-    let commaPos = -1;
-    let depth = 0;
-    for (let i = 0; i < inner.length; i++) {
-        const c = inner.charAt(i);
-        if (c === '(') depth++;
-        else if (c === ')') depth--;
-        else if (c === ',' && depth === 0) { commaPos = i; break; }
-    }
-
-    let name: string;
-    let fallback: string | undefined;
-    if (commaPos >= 0) {
-        name = inner.slice(0, commaPos).trim();
-        fallback = inner.slice(commaPos + 1).trim();
-    } else {
-        name = inner.trim();
-    }
-    if (name.startsWith('--')) name = name.slice(2);
-    if (!name) return value;
-
-    // Tier 1: runtime developer-set registry.
-    const reg = (globalThis as Record<string, unknown>).__pulpCssVars as
-        | Record<string, string> | undefined;
-    if (reg && typeof reg[name] === 'string' && reg[name]) {
-        return reg[name];
-    }
-
-    // Tier 2: theme.strings via bridge.
-    const getStr = (globalThis as Record<string, unknown>).getStringToken as
-        | ((n: string) => string) | undefined;
-    if (typeof getStr === 'function') {
-        const sv = getStr(name);
-        if (typeof sv === 'string' && sv) return sv;
-    }
-
-    // Tier 3: theme.dimensions via bridge (numeric tokens).
-    const getNum = (globalThis as Record<string, unknown>).getMotionToken as
-        | ((n: string) => number) | undefined;
-    if (typeof getNum === 'function') {
-        const nv = getNum(name);
-        if (typeof nv === 'number' && nv !== 0 && Number.isFinite(nv)) {
-            return String(nv);
-        }
-    }
-
-    // Tier 4: explicit fallback. Recurse so `var(--a, var(--b, 0))`
-    // walks the chain. Cap recursion via the input shrinking strictly
-    // on each step (we strip one var() wrapper per call).
-    if (fallback !== undefined && fallback.length > 0) {
-        return _resolveVar(fallback);
-    }
-
-    // Tier 5: surface the original unresolved string. Better than ""
-    // because downstream Skia / View code can at least log the unknown
-    // token name.
-    return value;
-}
-
-// pulp #1434 (batch 3) — translate CSS / React-Native fontWeight keyword
-// values to numeric weights before reaching the bridge. Mirrors the same
-// logic in the JS CSS shim (`web-compat-style-decl.js`). Numeric values
-// (`400`, `'500'`) flow through unchanged. The previous `Number(value)`
-// fallback returned NaN for keywords like `'bold'`, which the bridge
-// then coerced to 400 — silently mapping bold to normal. CSS spec:
-//   normal  → 400
-//   bold    → 700
-//   lighter → 300 (no font cascade in pulp; safe lower default)
-//   bolder  → 700 (no font cascade in pulp; "one step bolder than
-//                  normal" lands on bold)
-function _normalizeFontWeight(value: unknown): number {
-    if (typeof value === 'number') return value;
-    const s = String(value).trim().toLowerCase();
-    if (s === 'normal') return 400;
-    if (s === 'bold') return 700;
-    if (s === 'lighter') return 300;
-    if (s === 'bolder') return 700;
-    const n = Number(s);
-    return Number.isFinite(n) ? n : 400;
-}
-
-// pulp #1434 (Triage #15) — parse a CSS-spec single-shadow `box-shadow`
-// string. Mirrors the regex in `core/view/js/web-compat-style-decl.js`
-// (the DOM-lite path) so the @pulp/react path produces identical
-// dispatch shape.
-//
-// Wave 2 rn — added comma-separated multi-shadow support:
-// `_splitMultiShadow` walks the string respecting paren depth (so an
-// `rgba(0,0,0,0.3)` color literal inside one shadow doesn't get cut by
-// its internal commas). Each substring then flows through the existing
-// single-shadow parser; the prop-applier dispatches one setBoxShadow
-// per parsed shadow. This matches the CSS spec layering order where
-// the FIRST shadow paints on TOP (closest to the viewer); the bridge
-// keeps last-write-wins on the View slot today, so order matters at
-// paint time. Mock-bridge captures one call per shadow.
-//
-// Format (per shadow): `[inset] <dx>px <dy>px <blur>px [<spread>px] <color>`
-// Returns the parsed shape or null if the string doesn't match.
-interface _ParsedBoxShadow {
-    offsetX: number;
-    offsetY: number;
-    blur: number;
-    spread: number;
-    color: string;
-    inset: boolean;
-}
-function _parseBoxShadow(s: string): _ParsedBoxShadow | null {
-    let work = s.trim();
-    let inset = false;
-    if (/^inset\s+/i.test(work)) {
-        inset = true;
-        work = work.replace(/^inset\s+/i, '');
-    } else if (/\s+inset\s*$/i.test(work)) {
-        inset = true;
-        work = work.replace(/\s+inset\s*$/i, '');
-    }
-    // <dx>px <dy>px <blur>px [<spread>px] <color>
-    const m = work.match(/(-?[\d.]+)px\s+(-?[\d.]+)px\s+([\d.]+)px(?:\s+(-?[\d.]+)px)?\s+(.*)/);
-    if (!m) return null;
-    return {
-        offsetX: parseFloat(m[1]),
-        offsetY: parseFloat(m[2]),
-        blur: parseFloat(m[3]),
-        spread: parseFloat(m[4] ?? '0'),
-        color: m[5].trim(),
-        inset,
-    };
-}
-
-// Wave 2 rn — split a comma-separated multi-shadow string into individual
-// single-shadow substrings. Respects paren depth so commas inside a
-// `rgba(...)` color literal don't split a single shadow. Returns the
-// list of trimmed substrings (or `[s.trim()]` for a single shadow with
-// no top-level comma).
-function _splitMultiShadow(s: string): string[] {
-    const out: string[] = [];
-    let depth = 0;
-    let start = 0;
-    for (let i = 0; i < s.length; i++) {
-        const c = s[i];
-        if (c === '(') depth++;
-        else if (c === ')') { if (depth > 0) depth--; }
-        else if (c === ',' && depth === 0) {
-            out.push(s.slice(start, i).trim());
-            start = i + 1;
-        }
-    }
-    out.push(s.slice(start).trim());
-    return out.filter((x) => x.length > 0);
-}
-
-// pulp #1434 (Triage #9) — RN's `transform` is an array of single-property
-// objects (`[{translateX:10},{rotate:'45deg'},{scale:1.5}]`). Figma /
-// v0.dev / Claude Design exports emit this constantly. The bridge has
-// setTranslate / setRotation / setScale (uniform), so the array-walker
-// accumulates a per-render snapshot inside one pass and emits at most
-// three consolidated calls. Within-array merging means
-// `[{translateX:10},{translateY:20}]` produces ONE setTranslate(10,20)
-// instead of two clobbering ones (the latter would zero the unrelated
-// axis on each call). RN semantics also say each render's array is a
-// complete description — absent fields reset to identity, so we don't
-// carry state across renders.
-//
-// Bridge gaps (deferred — TODOs + follow-up issue):
-//   • setScale is uniform-only — independent scaleX/scaleY axes can't
-//     round-trip. We approximate: scale > scaleX > scaleY in priority,
-//     last-write-wins within the array; if scaleX≠scaleY we emit the
-//     last seen and document the limitation.
-//   • rotateX/rotateY/perspective/matrix3d — 3D / matrix transforms not
-//     modeled in pulp's 2D View (no perspective; rotation is Z-axis
-//     only). Silently dropped with TODO.
-//   • matrix(a b c d tx ty) — 2D affine; the CSS shim decomposes to
-//     translate + uniform-scale + rotate components. The @pulp/react
-//     RN array surface doesn't have a matrix entry today (RN spec:
-//     only translateX/Y, scale, scaleX/Y, rotate/Z, skewX/Y), so the
-//     walker just silently drops `matrix`/`matrix3d` for parity.
-//
-// Triage #9 fan-out (this PR) — `setSkew` is now a registered bridge
-// function (View::set_skew has existed since the 2D slot was added).
-// The walker dispatches skewX/skewY by accumulating both axes and
-// emitting one consolidated setSkew(id, x_deg, y_deg) call.
-interface _TransformSnapshot {
-    tx: number;
-    ty: number;
-    rotateDeg: number;
-    scale: number;
-    skewX: number;
-    skewY: number;
-    haveTranslate: boolean;
-    haveRotate: boolean;
-    haveScale: boolean;
-    haveSkew: boolean;
-}
-
-// Parse `'45deg'` / `'0.785rad'` / `45` (numeric) → degrees.
-function _parseAngleDegrees(v: unknown): number {
-    if (typeof v === 'number') return v;
-    const s = String(v).trim();
-    const m = s.match(/^(-?[\d.]+)\s*(deg|rad|turn|grad)?$/i);
-    if (!m) return 0;
-    const n = parseFloat(m[1]);
-    const unit = (m[2] || 'deg').toLowerCase();
-    if (unit === 'rad')  return n * (180 / Math.PI);
-    if (unit === 'turn') return n * 360;
-    if (unit === 'grad') return n * 0.9;
-    return n;
-}
-
-// Wave 2 rn — coerce a single CSS-length-like token to the value shape
-// the bridge's per-edge setFlex keys expect. Numbers pass through
-// numerically; percent strings (`'5%'`) pass through verbatim (the
-// bridge detects the suffix and routes to Yoga's percent path); plain
-// numeric strings become numbers; everything else falls back to 0.
-// Used by the padding shorthand fan-out where the bridge's `padding`
-// shorthand key only accepts a number.
-function _coerceLen(tok: string | number): number | string {
-    if (typeof tok === 'number') return tok;
-    const s = String(tok).trim();
-    if (s.endsWith('%')) return s; // forwarded verbatim → bridge percent path
-    const n = parseFloat(s);
-    return Number.isFinite(n) ? n : 0;
-}
-
-// Wave 2 rn — like `_coerceLen` but ALSO honors the `'auto'` keyword
-// (Yoga's YGNodeStyleSetMarginAuto — used for centering with
-// `marginLeft: 'auto'` + `marginRight: 'auto'`). Padding has no auto
-// equivalent in Yoga, so the regular `_coerceLen` is used there.
-function _coerceMarginLen(tok: string | number): number | string {
-    if (typeof tok === 'number') return tok;
-    const s = String(tok).trim();
-    if (s === 'auto') return 'auto';
-    if (s.endsWith('%')) return s;
-    const n = parseFloat(s);
-    return Number.isFinite(n) ? n : 0;
-}
-
-// Wave 2 rn — coerce a borderRadius value (uniform number OR RN
-// Fabric elliptical `{ x, y }` object) to the single number the
-// Skia paint backend currently honors. The RN Fabric form differs x
-// from y to draw an ellipse-shaped corner; the bridge / Skia path
-// today only takes a uniform radius, so we average x and y as the
-// closest visual approximation. True elliptical rendering needs the
-// paint side to call SkRRect::setRectXY; tracked as a deferred gap.
-function _coerceRadius(value: unknown): number | string {
-    if (typeof value === 'number') return value;
-    if (value != null && typeof value === 'object') {
-        const o = value as { x?: number; y?: number };
-        const x = typeof o.x === 'number' ? o.x : 0;
-        const y = typeof o.y === 'number' ? o.y : 0;
-        return (x + y) / 2;
-    }
-    if (typeof value === 'string') {
-        // pulp #1663 — preserve "%" suffix so the bridge can route to
-        // percent-aware paint-time resolution. Other unit suffixes (px,
-        // em, etc.) collapse to a plain number as before.
-        const trimmed = value.trim();
-        if (trimmed.endsWith('%')) {
-            const n = parseFloat(trimmed);
-            return Number.isFinite(n) ? `${n}%` : 0;
-        }
-        const n = parseFloat(trimmed);
-        return Number.isFinite(n) ? n : 0;
-    }
-    return 0;
-}
-
-// Wave 2 rn — resolve a `lineHeight` value with optional unitless
-// multiplier semantics. CSS spec: a unitless number on `line-height`
-// means "multiply by the element's font-size" (e.g.
-// `line-height: 1.5` with `font-size: 16` → 24px). We detect unitless
-// multipliers by treating values <= 8 as ratios (RN ports + design-tool
-// exports virtually always emit either `lineHeight: 24` (px) or
-// `lineHeight: 1.5` (multiplier); the threshold of 8 leaves room for
-// extremely large ratios while keeping common px values like 16 / 24 /
-// 32 unambiguous). Scaled by the live `fontSize` from props (defaults
-// to 14 — same Label default the bridge uses when no fontSize is
-// present). Numeric strings ("1.5") are treated like numbers; px-suffix
-// strings ("24px") strip the suffix and treat as absolute.
-function _resolveLineHeight(value: unknown, props: Record<string, unknown> | undefined): number {
-    let n: number;
-    if (typeof value === 'number') {
-        n = value;
-    } else {
-        const s = String(value).trim();
-        const sp = s.endsWith('px') ? s.slice(0, -2) : s;
-        n = parseFloat(sp);
-    }
-    if (!Number.isFinite(n)) return 0;
-    if (n > 0 && n <= 8) {
-        // Unitless multiplier — scale by current font-size.
-        const fs = (props && typeof props.fontSize === 'number')
-            ? (props.fontSize as number)
-            : 14;
-        return n * fs;
-    }
-    return n;
-}
-
-// pulp #1434 rn bridge-wires bundle — parse a CSS transform-origin
-// string into two fractional coordinates (0..1) the bridge expects.
-// Accepts `'center'`, `'left top'`, `'NN%'` percentages, and `'NNpx'`
-// pixel offsets (the latter assumed to be on a unit-bound View — so
-// values just clamp). Falls back to {0.5, 0.5} on unrecognized input.
-function _parseTransformOrigin(s: string): { x: number; y: number } {
-    const work = s.trim().toLowerCase();
-    if (work === 'center' || work === '') return { x: 0.5, y: 0.5 };
-    const tokens = work.split(/\s+/);
-    const tok2coord = (tok: string, axis: 'x' | 'y'): number => {
-        if (tok === 'center') return 0.5;
-        if (tok === 'left' || tok === 'top')   return 0.0;
-        if (tok === 'right' || tok === 'bottom') return 1.0;
-        if (tok.endsWith('%')) {
-            const n = parseFloat(tok.slice(0, -1));
-            return Number.isFinite(n) ? n / 100 : 0.5;
-        }
-        // Bare number / px — interpret as fractional 0..1 if <=1, else
-        // clamp to 1.0 (better than negative/over-1 garbage on the
-        // bridge side; full pixel resolution would need parent bounds).
-        const n = parseFloat(tok);
-        if (!Number.isFinite(n)) return 0.5;
-        return Math.max(0, Math.min(1, n));
-        void axis;
-    };
-    const x = tok2coord(tokens[0] ?? 'center', 'x');
-    const y = tok2coord(tokens[1] ?? tokens[0] ?? 'center', 'y');
-    return { x, y };
-}
-
-// Walk the RN-style transform array. Returns a snapshot with `have*`
-// flags so the caller can dispatch only the operations the user
-// actually specified (translate dispatch is gated on haveTranslate, etc.).
-function _walkTransformArray(arr: ReadonlyArray<unknown>): _TransformSnapshot {
-    const snap: _TransformSnapshot = {
-        tx: 0, ty: 0,
-        rotateDeg: 0,
-        scale: 1,
-        skewX: 0, skewY: 0,
-        haveTranslate: false,
-        haveRotate: false,
-        haveScale: false,
-        haveSkew: false,
-    };
-    for (const op of arr) {
-        if (op == null || typeof op !== 'object') continue;
-        const o = op as Record<string, unknown>;
-        const keys = Object.keys(o);
-        if (keys.length === 0) continue;
-        const k = keys[0];
-        const v = o[k];
-        switch (k) {
-            case 'translateX':
-                snap.tx = typeof v === 'number' ? v : parseFloat(String(v));
-                snap.haveTranslate = true;
-                break;
-            case 'translateY':
-                snap.ty = typeof v === 'number' ? v : parseFloat(String(v));
-                snap.haveTranslate = true;
-                break;
-            case 'rotate':
-            case 'rotateZ':
-                snap.rotateDeg = _parseAngleDegrees(v);
-                snap.haveRotate = true;
-                break;
-            case 'scale':
-                snap.scale = typeof v === 'number' ? v : parseFloat(String(v));
-                snap.haveScale = true;
-                break;
-            case 'scaleX':
-            case 'scaleY':
-                // Bridge has uniform setScale only; last-write-wins.
-                // pulp follow-up will add setScaleXY for independent axes.
-                snap.scale = typeof v === 'number' ? v : parseFloat(String(v));
-                snap.haveScale = true;
-                break;
-            // pulp #1434 Triage #9 fan-out — skewX / skewY now reach the
-            // bridge via the freshly-registered setSkew(id, x_deg, y_deg).
-            // Both axes accumulate independently; one consolidated call
-            // emits at dispatch time.
-            case 'skewX':
-                snap.skewX = _parseAngleDegrees(v);
-                snap.haveSkew = true;
-                break;
-            case 'skewY':
-                snap.skewY = _parseAngleDegrees(v);
-                snap.haveSkew = true;
-                break;
-            // 3D / matrix ops — not modeled in pulp's 2D View. Silently
-            // drop. pulp follow-up tracks if/when 3D transforms are
-            // introduced.
-            case 'rotateX':
-            case 'rotateY':
-            case 'perspective':
-            case 'matrix':
-                break;
-            default:
-                // Unknown op — silently drop (matches CSS shim tolerance).
-                break;
-        }
-    }
-    return snap;
-}
-
 function applyEventHandler(id: string, key: string, value: unknown): void {
     if (typeof value !== 'function') return;
     const eventName = eventNameFor(key);
@@ -628,6 +185,16 @@ function emitSvgLineGeometry(id: string, props: Record<string, unknown>): void {
 }
 
 /// Apply a single prop to its corresponding bridge setter.
+///
+/// P5-NEW-A — thin dispatcher over the per-domain handler modules.
+/// Each `applyXProp` returns true if it claimed the key. Domain keys
+/// are mutually exclusive (every prop belongs to exactly one domain),
+/// so the call order does not change which handler runs — it is
+/// byte-identical to the pre-split source-ordered switch. The
+/// type-dispatched widget/SVG props (`data` / `level` / `value` / `d`
+/// / `viewBox` / `fill` / `stroke` / `strokeWidth`) stay inline below
+/// because they route on `type`, not purely on `key`, and do not fit
+/// the layout/paint/typography/transform/events taxonomy.
 function applyOne(id: string, type: string, key: string, value: unknown, props?: Record<string, unknown>): void {
     if (value === undefined || value === null) {
         // No-op — Pulp has no "unset" for most setters; rely on React
@@ -655,1024 +222,19 @@ function applyOne(id: string, type: string, key: string, value: unknown, props?:
         }
     }
 
+    // Per-domain dispatch. The first handler that claims the key wins;
+    // since domain keys are disjoint this is equivalent to the
+    // pre-split single switch.
+    if (applyLayoutProp(id, key, value, props)) return;
+    if (applyPaintProp(id, key, value)) return;
+    if (applyTypographyProp(id, key, value, props)) return;
+    if (applyTransformProp(id, key, value)) return;
+    if (applyEventProp(id, key, value)) return;
+
+    // Type-dispatched widget / SVG props — these route on the widget
+    // `type`, not purely on the prop key, so they stay in the
+    // dispatcher rather than a key-keyed domain module.
     switch (key) {
-        // Flex / layout — all forwarded through setFlex
-        // pulp #1434 (rn NOT-IMPL bundle 1) — `direction` is overloaded:
-        //   • RN (and CSS spec) sense — writing direction: 'ltr' / 'rtl' /
-        //     'inherit' (RN spec also accepts 'auto' on iOS-classic). The
-        //     New Architecture surfaces this cross-platform.
-        //   • pulp historical sense — flexDirection alias: 'row' / 'col' /
-        //     'row-reverse' / 'column' / 'column-reverse'. Existing test
-        //     at prop-applier-direction.test.ts:60 pins this behavior.
-        // Disambiguate on value: writing-direction keywords route to
-        // setDirection; everything else falls through to setFlex(direction)
-        // for backward compat. `writingDirection` is preferred for new code
-        // (case below) but RN code commonly emits `direction: 'rtl'`.
-        case 'direction': {
-            const sval = String(value).trim().toLowerCase();
-            if (sval === 'ltr' || sval === 'rtl' || sval === 'inherit' || sval === 'auto') {
-                return call('setDirection', id, sval);
-            }
-            return call('setFlex', id, 'direction', value as string);
-        }
-        // pulp #108 — RN/React-style `flexDirection` (camelCase) is the
-        // canonical key in JSX. Without this case the prop fell through
-        // as unknown, leaving Yoga's column default in place and
-        // collapsing CSS-imported flex rows into vertical stacks. Maps
-        // to the same setFlex(direction, …) dispatch as the `direction`
-        // case for the flex-direction subset of values. Normalizes
-        // `col` / `col-reverse` aliases to `column` / `column-reverse`
-        // for the bridge's expected vocabulary.
-        case 'flexDirection': {
-            const sval = String(value).trim().toLowerCase();
-            const normalized = sval === 'col' ? 'column'
-                : sval === 'col-reverse' ? 'column-reverse'
-                : sval;
-            return call('setFlex', id, 'direction', normalized);
-        }
-        case 'gap':             return call('setFlex', id, 'gap', value as number);
-        case 'rowGap':          return call('setFlex', id, 'row_gap', value as number);
-        case 'columnGap':       return call('setFlex', id, 'column_gap', value as number);
-        // Wave 2 rn — `padding` shorthand accepts string forms (`'5%'`,
-        // `'10px 20px'`, etc.). The bridge `padding` shorthand key only
-        // takes a numeric value, so we fan out string values to the
-        // four per-edge keys (which DO accept `number | string` via
-        // setFlex(padding_top/...) and route through Yoga's
-        // YGNodeStyleSetPaddingPercent for `'5%'`). Numeric values
-        // continue to flow through the original shorthand path so we
-        // preserve the single-bridge-call shape for the common case.
-        case 'padding': {
-            if (typeof value === 'string') {
-                const tokens = value.trim().split(/\s+/);
-                const t = _coerceLen(tokens[0] ?? 0);
-                const r = _coerceLen(tokens[1] ?? tokens[0] ?? 0);
-                const b = _coerceLen(tokens[2] ?? tokens[0] ?? 0);
-                const l = _coerceLen(tokens[3] ?? tokens[1] ?? tokens[0] ?? 0);
-                call('setFlex', id, 'padding_top',    t);
-                call('setFlex', id, 'padding_right',  r);
-                call('setFlex', id, 'padding_bottom', b);
-                call('setFlex', id, 'padding_left',   l);
-                return;
-            }
-            return call('setFlex', id, 'padding', value as number);
-        }
-        // pulp #1434 (cross-surface mega-batch) — per-edge padding accepts
-        // either a number (px) or a percent string ('5%' → percent of
-        // parent main-axis size). Yoga padding does NOT support 'auto'.
-        case 'paddingTop':      return call('setFlex', id, 'padding_top',    value as number | string);
-        case 'paddingRight':    return call('setFlex', id, 'padding_right',  value as number | string);
-        case 'paddingBottom':   return call('setFlex', id, 'padding_bottom', value as number | string);
-        case 'paddingLeft':     return call('setFlex', id, 'padding_left',   value as number | string);
-        // Wave 2 rn — `margin` shorthand accepts string forms (`'5%'`,
-        // `'auto'`, `'10px auto'`, etc.). The bridge `margin` shorthand
-        // key only takes a numeric value, so we fan out string values
-        // to the four per-edge keys (which DO accept `number | string`
-        // including the `'auto'` keyword via Yoga's
-        // YGNodeStyleSetMarginAuto for centering). Numeric values
-        // continue through the original shorthand path so the common
-        // single-call case is preserved.
-        case 'margin': {
-            if (typeof value === 'string') {
-                const tokens = value.trim().split(/\s+/);
-                const t = _coerceMarginLen(tokens[0] ?? 0);
-                const r = _coerceMarginLen(tokens[1] ?? tokens[0] ?? 0);
-                const b = _coerceMarginLen(tokens[2] ?? tokens[0] ?? 0);
-                const l = _coerceMarginLen(tokens[3] ?? tokens[1] ?? tokens[0] ?? 0);
-                call('setFlex', id, 'margin_top',    t);
-                call('setFlex', id, 'margin_right',  r);
-                call('setFlex', id, 'margin_bottom', b);
-                call('setFlex', id, 'margin_left',   l);
-                return;
-            }
-            return call('setFlex', id, 'margin', value as number);
-        }
-        // pulp #1434 (cross-surface mega-batch) — per-edge margin accepts
-        // a number (px), percent string ('5%'), or the keyword 'auto'
-        // (Yoga's YGNodeStyleSetMarginAuto — used for centering with
-        // marginLeft:'auto' + marginRight:'auto').
-        case 'marginTop':       return call('setFlex', id, 'margin_top',    value as number | string);
-        case 'marginRight':     return call('setFlex', id, 'margin_right',  value as number | string);
-        case 'marginBottom':    return call('setFlex', id, 'margin_bottom', value as number | string);
-        case 'marginLeft':      return call('setFlex', id, 'margin_left',   value as number | string);
-        // pulp #1434 batch 4 — React Native shorthand aliases. RN code
-        // commonly writes `style={{ marginHorizontal: 8 }}` and expects
-        // it to fan out to marginLeft + marginRight on the underlying
-        // layout. Same pattern for marginVertical / paddingHorizontal /
-        // paddingVertical. We dispatch to the existing per-edge bridge
-        // setters so the value reaches the same FlexStyle slot whether
-        // it arrived through this alias or through the explicit edge
-        // prop. No FlexStyle field change required.
-        // pulp #1434 cross-surface mega-batch — RN aliases now forward
-        // percent strings (and 'auto' for margin) through the per-edge
-        // fan-out. The per-edge keys `margin_*` / `padding_*` accept
-        // number | string at the bridge boundary.
-        case 'marginHorizontal':
-            call('setFlex', id, 'margin_left',  value as number | string);
-            call('setFlex', id, 'margin_right', value as number | string);
-            return;
-        case 'marginVertical':
-            call('setFlex', id, 'margin_top',    value as number | string);
-            call('setFlex', id, 'margin_bottom', value as number | string);
-            return;
-        case 'paddingHorizontal':
-            call('setFlex', id, 'padding_left',  value as number | string);
-            call('setFlex', id, 'padding_right', value as number | string);
-            return;
-        case 'paddingVertical':
-            call('setFlex', id, 'padding_top',    value as number | string);
-            call('setFlex', id, 'padding_bottom', value as number | string);
-            return;
-        case 'flexGrow':        return call('setFlex', id, 'flex_grow', value as number);
-        case 'flexShrink':      return call('setFlex', id, 'flex_shrink', value as number);
-        // pulp #1434 (#1518) — RN-style `flex: <number>` shorthand.
-        // RN spec: `flex: positive` → `{flexGrow: n, flexShrink: 1, flexBasis: 0}`;
-        // `flex: 0` → no growth / no shrink at intrinsic basis;
-        // `flex: -1` (or any negative) → no growth, can shrink at auto basis.
-        // CSS spec is more nuanced (bare number is `flex: <n> 1 0`), but RN's
-        // narrow contract is what consumers passing JSX `flex={1}` expect;
-        // our adapter is RN-flavored so we honor RN semantics.
-        case 'flex': {
-            const n = value as number;
-            if (typeof n !== 'number' || !Number.isFinite(n)) return;
-            if (n > 0) {
-                call('setFlex', id, 'flex_grow', n);
-                call('setFlex', id, 'flex_shrink', 1);
-                call('setFlex', id, 'flex_basis', 0);
-            } else if (n === 0) {
-                call('setFlex', id, 'flex_grow', 0);
-                call('setFlex', id, 'flex_shrink', 0);
-                call('setFlex', id, 'flex_basis', 'auto');
-            } else {
-                call('setFlex', id, 'flex_grow', 0);
-                call('setFlex', id, 'flex_shrink', 1);
-                call('setFlex', id, 'flex_basis', 'auto');
-            }
-            return;
-        }
-        // pulp #1434 (rn batch C) — dimension keys forward
-        // `number | string` so the bridge sees `'50%'` / `'auto'`
-        // verbatim. Numeric values still flow through unchanged.
-        // The bridge's setFlex case for each key inspects the third
-        // arg as a string and detects '%' / 'auto' suffix; otherwise
-        // it falls back to the numeric path.
-        case 'flexBasis':       return call('setFlex', id, 'flex_basis', value as number | string);
-        // pulp #1434 Triage #14 — flexWrap accepts boolean (legacy
-        // true/false) or the CSS keyword strings (`"wrap"` /
-        // `"wrap-reverse"` / `"nowrap"`). Forward strings verbatim
-        // so the bridge can route wrap-reverse through Yoga's
-        // YGWrapWrapReverse.
-        case 'flexWrap': {
-            if (typeof value === 'string') {
-                return call('setFlex', id, 'flex_wrap', value);
-            }
-            return call('setFlex', id, 'flex_wrap', value ? 1 : 0);
-        }
-        case 'order':           return call('setFlex', id, 'order', value as number);
-        case 'width':           return call('setFlex', id, 'width', value as number | string);
-        case 'height':          return call('setFlex', id, 'height', value as number | string);
-        case 'minWidth':        return call('setFlex', id, 'min_width', value as number | string);
-        case 'minHeight':       return call('setFlex', id, 'min_height', value as number | string);
-        case 'maxWidth':        return call('setFlex', id, 'max_width', value as number | string);
-        case 'maxHeight':       return call('setFlex', id, 'max_height', value as number | string);
-        case 'alignItems':      return call('setFlex', id, 'align_items', value as string);
-        case 'alignSelf':       return call('setFlex', id, 'align_self', value as string);
-        // pulp #1434 (sub-agent #12 follow-up) — multi-line flex
-        // cross-axis distribution. Yoga supports it natively via
-        // YGNodeStyleSetAlignContent; the bridge accepts both bare
-        // (`start`/`end`) and prefixed (`flex-start`/`flex-end`)
-        // spellings plus the space-* values.
-        case 'alignContent':    return call('setFlex', id, 'align_content', value as string);
-        case 'justifyContent':  return call('setFlex', id, 'justify_content', value as string);
-        // pulp #1434 — aspectRatio routes through setFlex like the other
-        // flex props. Accepts a finite positive number (RN-style); strings
-        // ("16/9", "auto") are NOT accepted at the JSX surface — those
-        // belong to the CSS shim path (web-compat-style-decl.js). A value
-        // of 0 / NaN / undefined clears the slot on the bridge side.
-        case 'aspectRatio':     return call('setFlex', id, 'aspect_ratio', value as number);
-
-        // Visual style
-        // CSS `background` shorthand can carry color, gradient, or url. The
-        // C++ `setBackground` bridge fn (widget_bridge.cpp:3350) only parses
-        // a color token via `set_background_color(parseHexColor(...))`, so
-        // gradient strings collapse to a bogus color (often white). Caught
-        // by Spectr's 2026-05-11 live render — chip strip's
-        // `background: linear-gradient(to bottom, ...)` painted white, making
-        // the white SPECTR / ZOOMABLE FILTER BANK / axis labels invisible.
-        // Mirrors the same gradient-detection fix applied to backgroundImage
-        // below (Codex P1 on #1831).
-        case 'background': {
-            const sval = String(value);
-            if (/(linear|radial|conic)-gradient\(/i.test(sval)) {
-                return call('setBackgroundGradient', id, sval);
-            }
-            if (/^\s*url\s*\(/i.test(sval)) return;
-            // pulp #1899 (gap #3) — resolve var() for solid-color backgrounds
-            // (`background: var(--panel)`). Gradient strings keep their
-            // raw value because the gradient parser handles var() itself
-            // via the C++ shim.
-            return call('setBackground', id, _resolveVar(sval) as string);
-        }
-        case 'backgroundGradient': return call('setBackgroundGradient', id, value as string);
-        // CSS `background-image` longhand. The C++ `setBackground` bridge fn
-        // only parses a color token (widget_bridge.cpp:3350 →
-        // `set_background_color(parseHexColor(...))`), so we must detect
-        // gradient strings here and route to the gradient bridge instead;
-        // otherwise inputs like `linear-gradient(...)` collapse to a bogus
-        // color parse. `url(...)` has no image bridge today — drop quietly
-        // rather than corrupting the color slot. (Codex P1 on #1831.)
-        case 'backgroundImage': {
-            const sval = String(value);
-            if (/(linear|radial|conic)-gradient\(/i.test(sval)) {
-                return call('setBackgroundGradient', id, sval);
-            }
-            if (/^\s*url\s*\(/i.test(sval)) {
-                return;
-            }
-            // pulp #1899 (gap #3) — mirror `background` above.
-            return call('setBackground', id, _resolveVar(sval) as string);
-        }
-        // pulp #1517 — background sub-properties. The bridge stores the
-        // keyword on the View slot. Paint impact today is partial:
-        //   • `backgroundAttachment` — `scroll` is conformant; `fixed` /
-        //     `local` are noop (pulp doesn't model scroll contexts).
-        //   • `backgroundClip` — `text` is the interesting variant
-        //     (paint-time SkBlendMode::kSrcIn against text glyphs);
-        //     deferred to a future PR. Other values noop on solid bg.
-        //   • `backgroundOrigin` — relevant only for repeating gradients;
-        //     noop today.
-        case 'backgroundAttachment': return call('setBackgroundAttachment', id, value as string);
-        case 'backgroundClip':       return call('setBackgroundClip',       id, value as string);
-        case 'backgroundOrigin':     return call('setBackgroundOrigin',     id, value as string);
-        case 'border': {
-            const b = value as { color: string; width?: number; radius?: number };
-            return call('setBorder', id, b.color, b.width ?? 1, b.radius ?? 0);
-        }
-        // pulp #1027 (audit PR #1166 finding #4) — RN-style flat border props.
-        // These MUST route through the per-attribute bridge setters so a
-        // commitUpdate that touches only one of them preserves the others.
-        // Lowering them onto the unified `setBorder(id, color, width, radius)`
-        // would clobber the unset slots back to 0/empty.
-        case 'borderColor':  return call('setBorderColor', id, _resolveVar(value) as string);
-        case 'borderWidth':  return call('setBorderWidth', id, value as number);
-        // Wave 2 rn — `borderRadius` accepts the RN Fabric elliptical
-        // form `{ x, y }`. The Skia paint backend currently takes a
-        // single uniform radius per corner (no rrect ellipse axes), so
-        // we degrade the elliptical input by averaging x and y — the
-        // closest visual fidelity for the common Fabric usage where x
-        // and y differ only modestly. True elliptical rendering needs
-        // a paint-side rrect (Skia SkRRect::setRectXY) and remains a
-        // deferred gap. Numeric values flow through unchanged.
-        case 'borderRadius': return call('setBorderRadius', id, _coerceRadius(value));
-        // pulp #1434 Triage #10 — borderStyle keyword passes verbatim
-        // to setBorderStyle. Bridge maps to View::BorderStyle. Skia
-        // installs the dash effect for `dashed` / `dotted`; other
-        // named styles currently degrade to solid.
-        case 'borderStyle':  return call('setBorderStyle', id, value as string);
-        // pulp #1514 — list-style cluster. Pulp doesn't model
-        // <li>/<ul>/<ol> semantics, so the bridge stores the value
-        // verbatim on the View and a future paint pass renders the
-        // marker. Today the catalog is `partial` (stored, not
-        // painted). The shorthand `listStyle` parses on the JS side
-        // into the 3 longhands; consumers MAY emit any combo of
-        // type / position / image keywords (CSS spec: any order).
-        case 'listStyle': {
-            const sval = String(value).trim();
-            const tokens = sval.split(/\s+/);
-            const typeSet: Record<string, true> = {
-                none: true, disc: true, circle: true, square: true, decimal: true,
-            };
-            const posSet: Record<string, true> = { inside: true, outside: true };
-            let sawType = false, sawImage = false;
-            for (const tok of tokens) {
-                if (tok.indexOf('url(') === 0) {
-                    call('setListStyleImage', id, tok);
-                    sawImage = true;
-                } else if (posSet[tok]) {
-                    call('setListStylePosition', id, tok);
-                } else if (typeSet[tok]) {
-                    if (tok === 'none' && sawType && !sawImage) {
-                        call('setListStyleImage', id, 'none');
-                        sawImage = true;
-                    } else {
-                        call('setListStyleType', id, tok);
-                        sawType = true;
-                    }
-                }
-            }
-            return;
-        }
-        case 'listStyleType':     return call('setListStyleType', id, value as string);
-        case 'listStyleImage':    return call('setListStyleImage', id, value as string);
-        case 'listStylePosition': return call('setListStylePosition', id, value as string);
-        case 'borderTop':    { const b = value as { color: string; width: number }; return call('setBorderSide', id, 'top', b.width, b.color); }
-        case 'borderRight':  { const b = value as { color: string; width: number }; return call('setBorderSide', id, 'right', b.width, b.color); }
-        case 'borderBottom': { const b = value as { color: string; width: number }; return call('setBorderSide', id, 'bottom', b.width, b.color); }
-        case 'borderLeft':   { const b = value as { color: string; width: number }; return call('setBorderSide', id, 'left', b.width, b.color); }
-        // RN per-side flat props — route to the per-side bridge setters
-        // that already preserve the unrelated attribute (see widget_bridge
-        // applyBorderSide helper introduced in pulp #1026).
-        case 'borderTopColor':       return call('setBorderTopColor', id, _resolveVar(value) as string);
-        case 'borderRightColor':     return call('setBorderRightColor', id, _resolveVar(value) as string);
-        case 'borderBottomColor':    return call('setBorderBottomColor', id, _resolveVar(value) as string);
-        case 'borderLeftColor':      return call('setBorderLeftColor', id, _resolveVar(value) as string);
-        case 'borderTopWidth':       return call('setBorderTopWidth', id, value as number);
-        case 'borderRightWidth':     return call('setBorderRightWidth', id, value as number);
-        case 'borderBottomWidth':    return call('setBorderBottomWidth', id, value as number);
-        case 'borderLeftWidth':      return call('setBorderLeftWidth', id, value as number);
-        // Wave 2 rn — per-corner radii accept the RN Fabric elliptical
-        // `{ x, y }` form too (degraded to averaged uniform radius;
-        // see `borderRadius` above for the rrect rationale).
-        case 'borderTopLeftRadius':     return call('setBorderTopLeftRadius', id, _coerceRadius(value));
-        case 'borderTopRightRadius':    return call('setBorderTopRightRadius', id, _coerceRadius(value));
-        case 'borderBottomLeftRadius':  return call('setBorderBottomLeftRadius', id, _coerceRadius(value));
-        case 'borderBottomRightRadius': return call('setBorderBottomRightRadius', id, _coerceRadius(value));
-        // pulp #1519 — RN outline cluster. Paint-time ring drawn OUTSIDE
-        // the border-box (no Yoga layout impact). Each prop routes to its
-        // own per-attribute bridge fn so a JSX prop diff that touches one
-        // outline-* preserves the others. Style keyword set mirrors
-        // borderStyle (CSS spec is identical).
-        case 'outlineColor':  return call('setOutlineColor',  id, _resolveVar(value) as string);
-        case 'outlineOffset': return call('setOutlineOffset', id, value as number);
-        case 'outlineStyle':  return call('setOutlineStyle',  id, value as string);
-        case 'outlineWidth':  return call('setOutlineWidth',  id, value as number);
-        // CSS shorthand `outline: <width> <style> <color>` — mirror the
-        // parser already wired in web-compat-style-decl.js so JSX
-        // `style={{ outline: '1px solid red' }}` fans out to the three
-        // per-attribute setters identically to el.style.outline = '...'.
-        case 'outline': {
-            const m = String(value).match(/([\d.]+)px\s+(\w+)\s+(.+)/);
-            if (!m) return;
-            call('setOutlineWidth', id, parseFloat(m[1]));
-            call('setOutlineStyle', id, m[2]);
-            call('setOutlineColor', id, m[3].trim());
-            return;
-        }
-        case 'opacity':      return call('setOpacity', id, value as number);
-        case 'visible':      return call('setVisible', id, value as boolean);
-        // pulp #1434 (rn batch — Triage #12) — `display: 'flex' | 'none'`.
-        // RN exports + Figma / v0 / Claude Design HTML routinely emit
-        // `style={{ display: 'flex' }}` (the implicit default in pulp,
-        // but the prop-applier shouldn't drop it as unknown) or
-        // `style={{ display: 'none' }}` to hide a subtree. The yoga
-        // surface wired this for the CSS shim in #1422; this branch
-        // makes the same path reachable from RN-flavored JSX without a
-        // round-trip through the el.style proxy.
-        //
-        // 'none'  → setVisible(id, false). View::visible() is the
-        //           canonical "skip render + don't lay out" signal.
-        // 'flex'  → setVisible(id, true). Yoga's flex layout is pulp's
-        //           default; explicit 'flex' just confirms it.
-        // Anything else (block / inline-block / inline-flex / grid)
-        // is silently ignored at this layer — the CSS shim handles
-        // those for the el.style path; for RN consumers, the typical
-        // emission is just 'flex' / 'none'.
-        case 'display': {
-            const sval = String(value);
-            if (sval === 'none') return call('setVisible', id, false);
-            if (sval === 'flex') {
-                call('setVisible', id, true);
-                // pulp #1894 — CSS web-compat: `display: flex` defaults to
-                // `flex-direction: row`. Pulp's Yoga config defaults to
-                // FlexDirection::column (RN convention), so when neither
-                // flexDirection nor flexFlow-with-direction is also being
-                // set in the same prop batch we must explicitly emit row.
-                // Mirrors the CSS shim path at web-compat-style-decl.js
-                // (display:flex handler). Without this fallback the
-                // flat-prop path used by `style={{ display: 'flex' }}`
-                // JSX silently collapses every flex container to a
-                // vertical stack on import — first seen in Spectr's
-                // editor toolbar post-#1859 re-validation.
-                if (props) {
-                    // pulp #1898 (Codex review P2) — `direction` is a
-                    // third flex-direction alias in this prop-applier
-                    // (see the `case 'direction'` block above: a value
-                    // of 'row' / 'column' / 'row-reverse' / 'column-reverse'
-                    // routes to setFlex(direction); writing-direction
-                    // keywords like 'ltr' / 'rtl' / 'inherit' / 'auto'
-                    // route to setDirection instead). The default-row
-                    // suppression must therefore also honor a caller-
-                    // supplied `direction: 'column'`, otherwise the
-                    // explicit column is overwritten by the default row
-                    // when both `display: 'flex'` and `direction:` land
-                    // in the same prop batch. Restrict the check to
-                    // flex-axis values so `direction: 'rtl'` (writing
-                    // direction, not flex) still picks up the default.
-                    const hasDirectionFlexValue = (() => {
-                        if (!Object.prototype.hasOwnProperty.call(props, 'direction')) return false;
-                        const dv = (props as Record<string, unknown>)['direction'];
-                        if (typeof dv !== 'string') return false;
-                        const norm = dv.trim().toLowerCase();
-                        return norm === 'row' || norm === 'column'
-                            || norm === 'row-reverse' || norm === 'column-reverse'
-                            || norm === 'col' || norm === 'col-reverse';
-                    })();
-                    const hasFlexDirection =
-                        Object.prototype.hasOwnProperty.call(props, 'flexDirection') ||
-                        Object.prototype.hasOwnProperty.call(props, 'flex-direction') ||
-                        hasDirectionFlexValue;
-                    const ff = props.flexFlow;
-                    const flexFlowHasDirection =
-                        typeof ff === 'string' && /\b(row|column)\b/.test(ff);
-                    if (!hasFlexDirection && !flexFlowHasDirection) {
-                        call('setFlex', id, 'direction', 'row');
-                    }
-                }
-                return;
-            }
-            return; // unknown display value — leave View at current visibility
-        }
-        // pulp #1434 (Triage #15) — `boxShadow` accepts:
-        //  • `null` / `undefined` / `'none'` → clearBoxShadow
-        //  • String form (`'2px 4px 8px rgba(0,0,0,0.3)'` with optional
-        //    `inset`) — parsed inline below.
-        //  • Object form `{ offsetX, offsetY, blur?, spread?, color,
-        //    inset? }` — dispatched directly.
-        //
-        // Wave 2 rn — multi-shadow comma-separated string lists are now
-        // wired: we split on commas (respecting paren depth so a color
-        // literal like `rgba(0,0,0,0.3)` doesn't get cut), then dispatch
-        // one setBoxShadow per parsed shadow. CSS spec layers the first
-        // shadow on TOP (closest to viewer); the bridge applies them in
-        // dispatch order so paint order matches the input string.
-        case 'boxShadow': {
-            if (value == null || value === 'none' || value === '') {
-                return call('clearBoxShadow', id);
-            }
-            // Wave 4 rn — RN Fabric `BoxShadowValue[]` array form. Each
-            // element is `{ offsetX, offsetY, color, blurRadius?,
-            // spreadDistance?, inset? }` (RN spec field names; CSS
-            // box-shadow uses `blur` / `spread`). The bridge takes one
-            // dispatch per shadow; we clear first to avoid append-only
-            // accumulation across re-renders, then dispatch in order
-            // (CSS layers the first shadow on top — paint order matches
-            // input order, same as the multi-shadow string path above).
-            if (Array.isArray(value)) {
-                call('clearBoxShadow', id);
-                for (const s of value as ReadonlyArray<{
-                    offsetX: number; offsetY: number;
-                    blurRadius?: number; spreadDistance?: number;
-                    blur?: number; spread?: number;
-                    color: string; inset?: boolean;
-                }>) {
-                    if (!s) continue;
-                    const blur   = typeof s.blurRadius     === 'number' ? s.blurRadius
-                                 : typeof s.blur           === 'number' ? s.blur   : 4;
-                    const spread = typeof s.spreadDistance === 'number' ? s.spreadDistance
-                                 : typeof s.spread         === 'number' ? s.spread : 0;
-                    call('setBoxShadow', id, s.offsetX, s.offsetY, blur, spread,
-                         s.color, !!s.inset);
-                }
-                return;
-            }
-            if (typeof value === 'object') {
-                const s = value as { offsetX: number; offsetY: number; blur?: number; spread?: number; blurRadius?: number; spreadDistance?: number; color: string; inset?: boolean };
-                const blur   = typeof s.blurRadius     === 'number' ? s.blurRadius
-                             : typeof s.blur           === 'number' ? s.blur   : 4;
-                const spread = typeof s.spreadDistance === 'number' ? s.spreadDistance
-                             : typeof s.spread         === 'number' ? s.spread : 0;
-                return call('setBoxShadow', id, s.offsetX, s.offsetY,
-                            blur, spread,
-                            s.color, !!s.inset);
-            }
-            if (typeof value === 'string') {
-                const parts = _splitMultiShadow(value);
-                let emitted = 0;
-                for (const p of parts) {
-                    const parsed = _parseBoxShadow(p);
-                    if (parsed) {
-                        call('setBoxShadow', id, parsed.offsetX, parsed.offsetY,
-                             parsed.blur, parsed.spread, parsed.color, parsed.inset);
-                        emitted++;
-                    }
-                }
-                return; // unparseable shadows silently dropped; non-empty parts that fail just skip (matches CSS shim behavior)
-                void emitted;
-            }
-            return;
-        }
-        // pulp #1387 gap #1 — overflow was reachable via the DOM-lite
-        // path (web-compat-style-decl.js routes 'overflow' to setOverflow)
-        // but missing from the @pulp/react prop-applier, so JSX consumers
-        // setting `style={{ overflow: 'hidden' }}` silently dropped it.
-        // Spectr's dropdowns hit this — `width: 230 + overflow: hidden`
-        // on the dropdown row was being discarded, so the row grew to
-        // intrinsic content width and overflowed the container.
-        // Accepts the CSS keyword strings ('hidden' / 'visible' /
-        // 'scroll' / 'auto'); bridge maps to View::Overflow enum.
-        case 'overflow':     return call('setOverflow', id, value as string);
-        // CSS per-axis overflow. Pulp's View::Overflow is a single-axis
-        // paint-clip flag (same surface as `overflow`); both axes alias
-        // to the same setter. Already-supported in web-compat-style-decl.js
-        // via the same routing — wire the React-applier path so JSX
-        // `style={{ overflowY: 'scroll' }}` (common in scroll containers)
-        // doesn't silently drop.
-        case 'overflowX':
-        case 'overflowY':    return call('setOverflow', id, value as string);
-        // CSS `white-space` — text wrapping behavior (normal | nowrap | pre |
-        // pre-wrap). Bridge stores the slot on View; consumed by TextShaper
-        // when computing line breaks. Mirror el.style path so JSX
-        // `style={{ whiteSpace: 'nowrap' }}` doesn't silently drop.
-        case 'whiteSpace':   return call('setWhiteSpace', id, value as string);
-        // CSS `text-overflow` — clip | ellipsis. Bridge stores the slot for
-        // Label paint to consume.
-        case 'textOverflow': return call('setTextOverflow', id, value as string);
-        // CSS `backdrop-filter: blur(Npx)`. The bridge setter takes a numeric
-        // blur radius in px; mirror the parser in web-compat-style-decl.js
-        // (other filter functions are intentionally ignored — matches the
-        // `unsupportedValues: ["other filter functions"]` compat entry).
-        case 'backdropFilter': {
-            const sval = String(value).trim().toLowerCase();
-            if (sval === '' || sval === 'none') {
-                return call('setBackdropFilter', id, 0);
-            }
-            const bdm = sval.match(/blur\(\s*([\d.]+)\s*(px)?\s*\)/);
-            if (bdm) {
-                return call('setBackdropFilter', id, parseFloat(bdm[1]) || 0);
-            }
-            return;
-        }
-        // pulp #1516 — CSS box-sizing. Yoga 3.x honors the spec via
-        // YGNodeStyleSetBoxSizing; consumers passing JSX
-        // `boxSizing: 'border-box'` get the standard
-        // "padding+border are inside declared dimensions" behavior.
-        // Web designs almost universally reset to `border-box`.
-        case 'boxSizing':    return call('setBoxSizing', id, value as string);
-
-        // pulp #1434 Phase A2-3 — writing direction (RN ViewStyle uses
-        // `writingDirection` for this — CSS uses `direction`, but the
-        // pulp prop name `direction` already routes to FlexProps via
-        // setFlex above. The CSS-string-form `style.direction = 'rtl'`
-        // path goes through the el.style adapter's `direction` case
-        // which calls setDirection directly).
-        case 'writingDirection': return call('setDirection', id, value as string);
-
-        // pulp #1737 RN-OOS-fixup (catalog audit 2026-05-11) — these 4
-        // RN style props were classified `wontfix` in compat.json despite
-        // having fully-wired bridge fns AND css-side proof on the
-        // matching css/* surfaces (css/verticalAlign, css/textDecoration*
-        // all `supported`). One-line route closes the rn-side gap; the
-        // catalog flips from wontfix → supported on the same JSON edit.
-        //
-        // verticalAlign — RN's cross-platform vertical-align prop.
-        // textAlignVertical — RN-Android equivalent of verticalAlign;
-        //                     same setVerticalAlign target works for
-        //                     both since Pulp's Label::vertical_align_
-        //                     models what both CSS+RN-Android need.
-        // textDecorationColor / textDecorationStyle — text-decoration
-        //                     longhand setters; bridge already routes
-        //                     to Label::set_text_decoration_color /
-        //                     ::set_text_decoration_style.
-        case 'verticalAlign':         return call('setVerticalAlign', id, value as string);
-        case 'textAlignVertical':     return call('setVerticalAlign', id, value as string);
-        case 'textDecorationColor':   return call('setTextDecorationColor', id, _resolveVar(value) as string);
-        case 'textDecorationStyle':   return call('setTextDecorationStyle', id, value as string);
-
-        // pulp #1434 Phase A2-1 — CSS transitions. The bridge parses
-        // the full shorthand into a list of TransitionSpecs that the
-        // dispatcher (PR 2 of the ladder) consults when a property
-        // changes. Longhand fields apply uniformly across the parsed
-        // list (CSS spec semantics).
-        case 'transition':                return call('setTransition', id, value as string);
-        case 'transitionProperty':        return call('setTransitionProperty', id, value as string);
-        case 'transitionDuration': {
-            // Accept "200ms" / "0.3s" string OR plain number-of-seconds.
-            if (typeof value === 'number') {
-                return call('setTransitionDuration', id, value);
-            }
-            const s = String(value).trim();
-            const ms = s.endsWith('ms');
-            const n = parseFloat(s);
-            return call('setTransitionDuration', id, ms ? n / 1000 : n);
-        }
-        case 'transitionDelay': {
-            if (typeof value === 'number') {
-                return call('setTransitionDelay', id, value);
-            }
-            const s = String(value).trim();
-            const ms = s.endsWith('ms');
-            const n = parseFloat(s);
-            return call('setTransitionDelay', id, ms ? n / 1000 : n);
-        }
-        case 'transitionTimingFunction':  return call('setTransitionTimingFunction', id, value as string);
-
-        // pulp #1434 Phase A2-1 — CSS animations. animation-name
-        // resolves through the keyframes registry populated by
-        // defineKeyframes; PR 4 wires the playback driver. The
-        // shorthand path takes a single name + duration; longhand
-        // props can be split out by the host as needed.
-        case 'animationName':   return call('setAnimation', id, value as string, 1.0, 1, 'normal');
-        // Codex audit on pulp #1508 (P1): animationDuration was
-        // dispatched to setTransitionDuration here, which mutated
-        // *transition* timing on the same View instead of *animation*
-        // timing. Route through the legacy 2-arg setAnimation control-
-        // token form (`setAnimation(id, 'duration', seconds)`) so the
-        // bridge stages it on the View's pending-animation slot
-        // alongside animationName / animationDelay / etc.
-        case 'animationDuration': {
-            if (typeof value === 'number') {
-                return call('setAnimation', id, 'duration', value);
-            }
-            const s = String(value).trim();
-            const ms = s.endsWith('ms');
-            const n = parseFloat(s);
-            return call('setAnimation', id, 'duration', ms ? n / 1000 : n);
-        }
-        case 'animationDelay': {
-            if (typeof value === 'number') {
-                return call('setAnimation', id, 'delay', value);
-            }
-            const s = String(value).trim();
-            const ms = s.endsWith('ms');
-            const n = parseFloat(s);
-            return call('setAnimation', id, 'delay', ms ? n / 1000 : n);
-        }
-        case 'animationTimingFunction':
-            return call('setAnimation', id, 'easing', value as string);
-        case 'animationIterationCount':
-            return call('setAnimation', id, 'iterations',
-                value === 'infinite' ? -1 : (typeof value === 'number' ? value : parseFloat(String(value)) || 1));
-        case 'animationDirection':
-            return call('setAnimation', id, 'direction', value as string);
-        case 'animationFillMode':
-            return call('setAnimation', id, 'fill', value as string);
-        // pulp #1434 Wave 3 css.3 — animation-play-state. Routes
-        // through the legacy 2-arg setAnimation control-token form so
-        // the bridge stores the keyword on View::animation_play_state_;
-        // View::tick_animations honors `paused` by skipping the
-        // timeline advance (web spec semantic).
-        case 'animationPlayState':
-            return call('setAnimation', id, 'play_state', value as string);
-
-        // pulp #1434 rn logical-edge bundle (sub-agent #27 finding) —
-        // RN's CSS-spec-equivalent logical-flow props. LTR-only fast
-        // path: Start → Left, End → Right, inset shorthand expands to
-        // top/right/bottom/left, insetBlock → top+bottom,
-        // insetInline → left+right (LTR). True RTL bidi requires a
-        // future direction system — tracked as a separate big project.
-        // The 11 entries here close the missing-on-rn gap with the
-        // honest LTR-only caveat documented in the catalog.
-        case 'marginStart': {
-            return call('setFlex', id, 'margin_left', value as number | string);
-        }
-        case 'marginEnd': {
-            return call('setFlex', id, 'margin_right', value as number | string);
-        }
-        case 'paddingStart': {
-            return call('setFlex', id, 'padding_left', value as number | string);
-        }
-        case 'paddingEnd': {
-            return call('setFlex', id, 'padding_right', value as number | string);
-        }
-        case 'borderStartWidth': {
-            // Routes to the per-side bridge setter that preserves the
-            // unrelated attribute (color) — same pattern as borderLeftWidth.
-            return call('setBorderLeftWidth', id, value as number);
-        }
-        case 'borderEndWidth': {
-            return call('setBorderRightWidth', id, value as number);
-        }
-        case 'start': {
-            return call('setLeft', id, value as number | string);
-        }
-        case 'end': {
-            return call('setRight', id, value as number | string);
-        }
-        // CSS `inset` shorthand: 1 / 2 / 3 / 4 values fan out to
-        // top / right / bottom / left (CSS spec — same expansion as
-        // `margin` / `padding` shorthands). Numeric or percent strings.
-        case 'inset': {
-            const v = value as number | string;
-            if (typeof v === 'number') {
-                call('setTop',    id, v);
-                call('setRight',  id, v);
-                call('setBottom', id, v);
-                call('setLeft',   id, v);
-                return;
-            }
-            const tokens = String(v).trim().split(/\s+/);
-            const t = tokens[0] ?? 0;
-            const r = tokens[1] ?? t;
-            const b = tokens[2] ?? t;
-            const l = tokens[3] ?? r;
-            // Each token may be a number or a percent string — forward
-            // verbatim so the bridge can route through Yoga's
-            // YGNodeStyleSetPositionPercent path for percent values.
-            const coerce = (tok: string | number) => {
-                if (typeof tok === 'number') return tok;
-                if (tok.endsWith('%')) return tok;
-                const n = parseFloat(tok);
-                return Number.isFinite(n) ? n : tok;
-            };
-            call('setTop',    id, coerce(t));
-            call('setRight',  id, coerce(r));
-            call('setBottom', id, coerce(b));
-            call('setLeft',   id, coerce(l));
-            return;
-        }
-        case 'insetBlock': {
-            // CSS insetBlock shorthand → top + bottom.
-            const v = value as number | string;
-            call('setTop',    id, v);
-            call('setBottom', id, v);
-            return;
-        }
-        case 'insetInline': {
-            // CSS insetInline shorthand → left + right (LTR).
-            const v = value as number | string;
-            call('setLeft',  id, v);
-            call('setRight', id, v);
-            return;
-        }
-
-        // pulp #1434 Phase A2-2 — CSS Grid surface. Forwards each
-        // property verbatim to setGrid; the C++ bridge handles
-        // template-track parsing, named-area parsing, and the
-        // grid-area shorthand (named token vs `row / col / row / col`
-        // numeric form).
-        case 'gridTemplateColumns': return call('setGrid', id, 'template_columns', value as string);
-        case 'gridTemplateRows':    return call('setGrid', id, 'template_rows',    value as string);
-        case 'gridTemplateAreas':   return call('setGrid', id, 'template_areas',   value as string);
-        case 'gridAutoColumns':     return call('setGrid', id, 'auto_columns',     value as string);
-        case 'gridAutoRows':        return call('setGrid', id, 'auto_rows',        value as string);
-        case 'gridAutoFlow':        return call('setGrid', id, 'auto_flow',        value as string);
-        case 'gridArea':            return call('setGrid', id, 'grid_area',        value as string);
-        case 'gridColumn': {
-            const parts = String(value).split('/').map((s) => parseInt(s.trim(), 10));
-            if (parts[0]) call('setGrid', id, 'column_start', parts[0]);
-            if (parts[1]) call('setGrid', id, 'column_end', parts[1]);
-            return;
-        }
-        case 'gridRow': {
-            const parts = String(value).split('/').map((s) => parseInt(s.trim(), 10));
-            if (parts[0]) call('setGrid', id, 'row_start', parts[0]);
-            if (parts[1]) call('setGrid', id, 'row_end', parts[1]);
-            return;
-        }
-        case 'gridColumnStart': return call('setGrid', id, 'column_start', value as number);
-        case 'gridColumnEnd':   return call('setGrid', id, 'column_end',   value as number);
-        case 'gridRowStart':    return call('setGrid', id, 'row_start',    value as number);
-        case 'gridRowEnd':      return call('setGrid', id, 'row_end',      value as number);
-        case 'gridGap':         return call('setGrid', id, 'gap',          value as number);
-        case 'gridColumnGap':   return call('setGrid', id, 'column_gap',   value as number);
-        case 'gridRowGap':      return call('setGrid', id, 'row_gap',      value as number);
-
-        // pulp #1434 rn bridge-wires bundle (sub-agent #27 finding) —
-        // 7 RN-style props that already had C++ bridge fns registered
-        // but no `@pulp/react` prop-applier dispatch. Each forwards
-        // the keyword / string straight through to the matching setter.
-        case 'backfaceVisibility': return call('setBackfaceVisibility', id, value as string);
-        case 'cursor':             return call('setCursor', id, value as string);
-        case 'filter':             return call('setFilter', id, value as string);
-        // pulp #1515 — CSS clip-path / mask cluster. The bridge fns
-        // store the value on the View; Skia paint side honors
-        // `clip-path: path("...")` via SkPath::FromSVGString. Other
-        // clip-path forms (URL refs, named shapes) and mask painting
-        // (saveLayer + SkBlendMode::kDstIn) are deferred. The
-        // prop-applier dispatches verbatim — keyword normalization
-        // and shorthand expansion live in the bridge / JS shim.
-        case 'clipPath':           return call('setClipPath', id, value as string);
-        case 'mask':               return call('setMask', id, value as string);
-        case 'maskImage':          return call('setMaskImage', id, value as string);
-        case 'maskSize':           return call('setMaskSize', id, value as string);
-        // CSS `appearance` — Pulp paints all widgets custom, so this
-        // is observably storage-only. Slot exists for round-trip.
-        case 'appearance':         return call('setAppearance', id, value as string);
-        // CSS `object-fit` / `object-position` — image fitting.
-        // Storage-only today; ImageView paint follow-up.
-        case 'objectFit':          return call('setObjectFit', id, value as string);
-        case 'objectPosition':     return call('setObjectPosition', id, value as string);
-        // pulp #1549 — RN `mixBlendMode` (RN 0.76 New Architecture).
-        // Forwards the W3C blend-mode keyword string to
-        // `setMixBlendMode`; the bridge keyword→enum table lives at
-        // widget_bridge.cpp::setMixBlendMode.
-        case 'mixBlendMode':       return call('setMixBlendMode', id, value as string);
-        case 'pointerEvents':      return call('setPointerEvents', id, value as string);
-        case 'textTransform':      return call('setTextTransform', id, value as string);
-        case 'userSelect':         return call('setUserSelect', id, value as string);
-        // transformOrigin accepts CSS strings of the form `'NN% NN%'`,
-        // `'NNpx NNpx'`, `'center'`, or two keyword tokens. The bridge
-        // wants two numeric fractions (0..1). Defaults to 0.5/0.5
-        // (center) when a token is unrecognized — matches CSS default.
-        case 'transformOrigin': {
-            const parsed = _parseTransformOrigin(String(value ?? 'center'));
-            return call('setTransformOrigin', id, parsed.x, parsed.y);
-        }
-
-        // pulp #1434 Triage #9 — RN array transform.
-        // RN's transform is an array of single-property objects:
-        //   transform: [
-        //     { translateX: 10 }, { translateY: 20 },
-        //     { rotate: '45deg' }, { scale: 1.5 },
-        //   ]
-        // Walk-once accumulates the snapshot in one pass (so
-        // {translateX:10} and {translateY:20} as separate entries
-        // produce ONE setTranslate(10,20), not two clobbering calls),
-        // then emits only the operations the user specified.
-        // Within-array semantics: each render's array is a complete
-        // description; absent fields reset to identity. No cross-
-        // render state is maintained — passing `transform: undefined`
-        // (or removing the prop) goes through the standard prop-
-        // removal path and resets translate/rotate/scale on the next
-        // re-render that includes the prop.
-        case 'transform': {
-            if (value == null) return;
-            if (!Array.isArray(value)) return;  // CSS string form deferred
-            const snap = _walkTransformArray(value as ReadonlyArray<unknown>);
-            if (snap.haveTranslate) call('setTranslate', id, snap.tx, snap.ty);
-            if (snap.haveRotate)    call('setRotation', id, snap.rotateDeg);
-            if (snap.haveScale)     call('setScale', id, snap.scale);
-            // pulp #1434 Triage #9 fan-out — setSkew is now a registered
-            // bridge fn; emit one consolidated call that captures both
-            // axes accumulated in the walker.
-            if (snap.haveSkew)      call('setSkew', id, snap.skewX, snap.skewY);
-            return;
-        }
-
-        // CSS-style positioning (pulp #779 follow-up; matches setPosition
-        // + setTop/setLeft/setRight/setBottom on the bridge).
-        // pulp #1434 batch 6 — top/right/bottom/left accept either a number
-        // ('50' → px) or a percent string ('50%' → percent of parent).
-        // Mirrors PR #1426 (width/height percent) for the four View
-        // positional fields. Figma absolute-positioned overlays, v0.dev
-        // hero anchors, and Claude Design sticky elements all emit
-        // `top:'50%'` etc. routinely; without percent forwarding the
-        // layout collapses to numeric 0 silently. The bridge inspects
-        // arg index 1 as a string, detects the '%' suffix, and routes to
-        // Yoga's YGNodeStyleSetPositionPercent path via View::top_unit_.
-        case 'position':     return call('setPosition', id, value as string);
-        case 'top':          return call('setTop', id, value as number | string);
-        case 'left':         return call('setLeft', id, value as number | string);
-        case 'right':        return call('setRight', id, value as number | string);
-        case 'bottom':       return call('setBottom', id, value as number | string);
-        case 'zIndex':       return call('setZIndex', id, value as number);
-
-        // Text
-        case 'text':            return call('setText', id, String(value));
-        // CSS-canonical `color` aliases RN-canonical `textColor`. Imported
-        // React designs (JSX `style={{ color: '...' }}`) silently dropped
-        // before this — bridge has `setTextColor`, the dispatch case was
-        // missing the alias.
-        case 'color':
-        case 'textColor':       return call('setTextColor', id, _resolveVar(value) as string);
-        // pulp #1434 — widen to include `'auto'` and `'justify'` (CSS /
-        // RN canonical). `'auto'` is writing-direction-relative
-        // (LTR-only today, degrades to `'left'`); `'justify'` flows to
-        // canvas TextAlign::justify (SkParagraph kJustify wiring is a
-        // follow-up — backends approximate as left for now).
-        case 'textAlign':       return call('setTextAlign', id, value as string);
-        // Typography — Label widgets honor these via setX bridge fns.
-        // pulp #1434 Phase A2-5 — fontFamily IS now dispatched. The
-        // bridge picks the first non-empty family from a comma-
-        // separated CSS list and stores it on the Label or on the
-        // owning View's `inheritable_font_family_` slot for container
-        // cascade. The whole-list fallback chain (full font-stack
-        // resolution) still depends on SkFontMgr registration in
-        // pulp #932 — until that lands, families that aren't already
-        // registered with Skia fall through to the platform default.
-        // Wiring is independent: when #932 lands, no consumer change
-        // is needed — the registry just resolves the same name.
-        // pulp #1899 (gap #3) — resolve `var(--mono)` before forwarding.
-        // The bridge expects a real family name; the literal "var(--mono)"
-        // gives Skia's font matcher nothing to match against and silently
-        // falls back to a proportional sans (e.g. Spectr top-bar labels
-        // rendered in the wrong typeface AND inside an opacity layer that
-        // degraded the LCD AA — both fixed in this change).
-        case 'fontFamily':      return call('setFontFamily', id, _resolveVar(value) as string);
-        case 'fontSize':        return call('setFontSize', id, value as number);
-        case 'fontWeight':      return call('setFontWeight', id, _normalizeFontWeight(value));
-        case 'fontStyle':       return call('setFontStyle', id, value as string);
-        case 'letterSpacing':   return call('setLetterSpacing', id, value as number);
-        // Wave 2 rn — `lineHeight` accepts CSS unitless-multiplier
-        // semantics. A value `<= 8` is treated as a multiplier of the
-        // current `fontSize` from props (defaults to 14 when absent —
-        // matches the Label default). Larger values flow through as
-        // absolute pixels (the existing path). Px-suffix strings strip
-        // the suffix and pass through as absolute too. The bridge
-        // setter signature is unchanged — it always sees a number.
-        case 'lineHeight':      return call('setLineHeight', id, _resolveLineHeight(value, props));
-        // pulp #1552 — line-clamp + webkit-line-clamp + background-repeat
-        // wiring. Both line-clamp keys funnel through the same setter
-        // (shared CSS shim case + RN-style alias). 0 / non-finite clears
-        // the slot. backgroundRepeat is storage-only at the View layer
-        // (paint-time honoring is a follow-up for url() / repeating
-        // gradient backgrounds).
-        case 'lineClamp':
-        case 'webkitLineClamp': {
-            const n = typeof value === 'number' ? value : parseInt(String(value), 10);
-            return call('setLineClamp', id, Number.isFinite(n) ? n : 0);
-        }
-        case 'backgroundRepeat': return call('setBackgroundRepeat', id, value as string);
-
-        // pulp #1434 (rn NOT-IMPL bundle 1) — RN's `textDecorationLine`
-        // is the spec-aligned name; pulp's bridge uses `setTextDecoration`
-        // (single-keyword form). RN allows `'underline line-through'` as
-        // a compound — pass through verbatim; the bridge's keyword table
-        // currently honors single-keyword `'none' / 'underline' /
-        // 'line-through' / 'overline'`. Compound rendering is the same
-        // partial gap as css/textDecoration's "single-keyword only" note.
-        case 'textDecorationLine': return call('setTextDecoration', id, value as string);
-        // pulp #1434 (rn NOT-IMPL bundle 1) — RN textShadow cluster.
-        // The CSS shim (`web-compat-style-decl.js`) calls `setTextShadow`
-        // defensively (`if (typeof setTextShadow === 'function')`); the
-        // bridge does NOT yet register that fn, so paint is a no-op
-        // today. Wire the @pulp/react surface here so when the bridge
-        // gains the registration (planned slot, see #1548 feature
-        // branch) every consumer flips on without a JSX-side change.
-        // Each per-attribute setter writes ONE slot in isolation so a
-        // diff that touches one prop doesn't clobber the others.
-        case 'textShadowColor':  return call('setTextShadowColor',  id, _resolveVar(value) as string);
-        case 'textShadowOffset': {
-            // RN spec: `{ width, height }` (number / number).
-            const o = value as { width?: number; height?: number };
-            const dx = typeof o?.width  === 'number' ? o.width  : 0;
-            const dy = typeof o?.height === 'number' ? o.height : 0;
-            return call('setTextShadowOffset', id, dx, dy);
-        }
-        case 'textShadowRadius': return call('setTextShadowRadius', id, value as number);
-
-        // pulp #1737 RN-OOS-fixup (audit 2026-05-11) — RN iOS-legacy
-        // box-shadow longhand. Modern RN code uses `boxShadow` (CSS
-        // shorthand) which Pulp fully supports, but upstream RN still
-        // accepts shadowColor / shadowOffset / shadowOpacity /
-        // shadowRadius as cross-platform-ish style props (originally
-        // iOS-only, but Pulp implements them cross-platform via the
-        // unified BoxShadow struct on View). Mirrors the per-attribute
-        // textShadow* fan-out above. Each per-attribute setter writes
-        // ONE slot of View::shadow_ in isolation so a JSX diff that
-        // touches one prop doesn't clobber the others.
-        case 'shadowColor':   return call('setShadowColor',   id, _resolveVar(value) as string);
-        case 'shadowOffset': {
-            // RN spec: `{ width, height }` (number / number).
-            const o = value as { width?: number; height?: number };
-            const dx = typeof o?.width  === 'number' ? o.width  : 0;
-            const dy = typeof o?.height === 'number' ? o.height : 0;
-            return call('setShadowOffset', id, dx, dy);
-        }
-        case 'shadowOpacity': return call('setShadowOpacity', id, value as number);
-        case 'shadowRadius':  return call('setShadowRadius',  id, value as number);
-
-        // pulp #1737 RN-OOS-fixup final sweep — RN's Android-only
-        // `elevation` (Material Design 0..24dp). The bridge shim
-        // translates the elevation value to a Material-approximated
-        // single-shadow BoxShadow so consumers shipping unchanged
-        // RN-Android styles render a visible shadow on every Pulp
-        // platform. Upstream RN ignores `elevation` on iOS entirely;
-        // Pulp's cross-platform translation is a strict improvement.
-        case 'elevation':     return call('setElevation', id, value as number);
-
-        // pulp #1737 RN-OOS-fixup final sweep — RN's Android-only
-        // `includeFontPadding`. Pulp's text-shaping doesn't add
-        // Android-vestigial vertical glyph padding regardless of this
-        // value, so the bridge fn accepts the keyword + stores it on
-        // a View slot (round-trip), and the paint pipeline ignores it.
-        // Authors setting `false` (the common case — remove Android
-        // padding) get what they want by default; authors setting
-        // `true` get the same tight-baseline behavior (Pulp can't add
-        // padding it doesn't model). Catalog status is `supported` as
-        // an honest CSS-subset claim (same pattern as overscroll-behavior).
-        case 'includeFontPadding': return call('setIncludeFontPadding', id, Boolean(value));
-
-        // pulp #1737 RN-OOS-fixup (#1812) — RN's `borderCurve` corner
-        // shape. `circular` (default) keeps Pulp's standard rounded
-        // corner; `continuous` switches the View paint to the iOS-
-        // style squircle approximation (super-ellipse path). Authors
-        // shipping iOS-aesthetic designs with `borderCurve: 'continuous'`
-        // now get the visible squircle on every Pulp platform.
-        case 'borderCurve':       return call('setBorderCurve', id, value as string);
-
-        // pulp #1737 RN-OOS-fixup (final round) — RN's `isolation`.
-        // Pulp's per-View paint model is structurally isolated by
-        // default (per-View save_layer_with_blend composition + paint-
-        // order-scoped z-index), matching the dominant author intent
-        // of `isolation: isolate`. Honest CSS-subset claim: both
-        // keywords round-trip; behavior matches `isolate` regardless.
-        case 'isolation':         return call('setIsolation', id, value as string);
-
-        // pulp #1434 (rn NOT-IMPL bundle 1) — RN's `experimental_backgroundImage`
-        // (New Architecture only) accepts a CSS gradient string. Route
-        // through the existing setBackgroundGradient bridge fn — same
-        // shape, same parser. RN also allows an array-of-objects gradient
-        // form on Fabric; that shape is NOT supported here (gradient
-        // strings only). Catalog flips missing → partial.
-        case 'experimental_backgroundImage':
-            return call('setBackgroundGradient', id, value as string);
-
-        // pulp #1434 (rn NOT-IMPL bundle 1) — RN's `fontVariant` is a
-        // string-array of OpenType feature tokens (`['small-caps',
-        // 'tabular-nums', ...]`). True implementation requires HarfBuzz
-        // hb_feature_t setting on the shape pass — outside the scope of
-        // this prop-applier wiring. Forward as a comma-joined string to
-        // a bridge fn name reserved for the future paint-time impl;
-        // until the bridge registers `setFontVariant`, the dispatch is
-        // a no-op (the `call` helper silently skips unregistered names).
-        // Catalog flips missing → partial with the gotcha documented.
-        case 'fontVariant': {
-            const joined = Array.isArray(value)
-                ? (value as string[]).join(',')
-                : String(value);
-            return call('setFontVariant', id, joined);
-        }
-
         // Widget-specific data
         case 'data':
             // <Spectrum data={...}> + <Waveform data={...}> share the prop
@@ -1690,45 +252,6 @@ function applyOne(id: string, type: string, key: string, value: unknown, props?:
             if (type === 'Progress')   return call('setProgress', id, value as number);
             if (type === 'Meter')      return call('setMeterLevel', id, value as number);
             return call('setValue', id, value as number);
-
-        // pulp #1148 — generalized overlay-click routing. `overlay={true}`
-        // claims the view as the active click-eligible overlay so React
-        // popovers built on `<View position="absolute">` receive clicks
-        // even though hit_test would otherwise resolve to a sibling. The
-        // matching releaseOverlay is emitted by applyChangedProps when
-        // the prop flips off, and by detach() at unmount.
-        case 'overlay':
-            if (value) return call('claimOverlay', id);
-            return call('releaseOverlay', id);
-
-        // pulp ARIA modal/popup auto-overlay — UX best-practice default.
-        // When the JSX declares an ARIA role that semantically IS a
-        // dismissable overlay (`role="dialog" | "alertdialog" | "menu" |
-        // "listbox"`) or sets `aria-modal="true"`, claim the overlay so
-        // Esc-dismiss + outside-click routing fire automatically. Pre-fix,
-        // every consumer (Spectr's dom-adapter, etc.) had to opt in by
-        // mirroring a position:absolute heuristic, which missed inset:0
-        // full-screen modal backdrops (the most common modal pattern) and
-        // every dropdown/menu authored without explicit positioning.
-        //
-        // Override semantics: an explicit `overlay={false}` still wins
-        // because applyChangedProps emits that case AFTER the role case
-        // (object iteration order is insertion order, and JSX collects
-        // props left-to-right; `overlay` typically appears after `role`).
-        // For defensive parity, an explicit overlay={true} is a no-op on
-        // top of the auto-claim (idempotent on the bridge side).
-        case 'role': {
-            const r = typeof value === 'string' ? value.toLowerCase() : '';
-            if (r === 'dialog' || r === 'alertdialog' || r === 'menu' || r === 'listbox') {
-                return call('claimOverlay', id);
-            }
-            return;
-        }
-        case 'aria-modal': {
-            const truthy = value === true || value === 'true' || value === '';
-            if (truthy) return call('claimOverlay', id);
-            return;
-        }
 
         // SvgPath (pulp #994) — wires the SvgPathWidget bridge surface
         // (createSvgPath / setSvgPath / setSvgViewBox / setSvgFill /
